Changed GUI appearance
